+++ ezwow/__init__.py
@@ -2,5 +2,5 @@
 
 from ezwow._entry import main
 
-__version__: str = "2.1.0"
+__version__: str = "2.2.0"
 __all__ = ["__version__", "main"]